Signal reader and listener
Introduces a signal reader to get digital
inputs from Arduino level switches. The
switch input is fed into the app through
the serial port, in the form:

[pin] (1 byte) [value] (1 byte)

These values are fed to a delegate that can
respond. Temporarily implemented changing
the color of the LED screen each time the
switch on pin 2 changes.