Audio stem integration with perfect sync
Integrate the audio stems so that each color/body-part,
when selected, plays its stem, and when all selected
pins are of the same color, the drums start as well.

It was extremely important to achieve perfect sync when
playing the various stems on top of one another. To
achieve this:

1. We start the sequence from the renderig thread of the
countdown screen, which serves also as the master timer
for the animations. This prevents issues where the dispatch
to the main thread was throwing off sync between the
animation and sound.

2. We use AVAudioEngine and AVAudioPlayerNode, which
fixed start times and preloading, to ensure perfect sync
between the various tracks.

This commit also includes an attempt at perfect sync
that was implemented earlier, called volume sync. In that
method, all tracks (all 24!) were played at all times, and
only the volumes of each track were modified. This method
will be removed shortly as its less efficient than the new
method.