Aubio-based beat detection
Integrate the Aubio library and microphone
recording to detect the beat of live music
in real-time. Convert the detected BPM into
an 8-beat period to make the characters
dance at that rate.

This is an experimental (fun) alternative mode
of interaction for the characters.

To enable, CCBeatDetection should be set in
the CCDanceController, and kMuteAudio should
be set to true.